Make Home tab as current weather screen
