+++ OpenWeather/App/Home/HomeViewModel.swift
@@ -2,13 +2,13 @@
 //  OpenWeather template generated by OpenBytes on 15/03/2023.
 //
 // Created by Ahmed Shendy.
-//  WeatherDetailsViewModel.swift
+//  HomeViewModel.swift
 //
 
 import Foundation
 import CoreLocation
 
-final class WeatherDetailsViewModel: ObservableObject, ErrorHandling {
+final class HomeViewModel: ObservableObject, ErrorHandling {
     private var weatherProviding: WeatherProviding
     private var locationProviding: LocationProviding
     private var task: Task<Void, Never>?
